Longer max-age
There's no longer any reason why this can't be cached for a longer interval. This change increases our LB cache hit rate from about 90% to 93-94%, which is a 30-40% reduction in load on the backend servers.

Still don't want to go *too* high, as it will start to elongate the time between when a release is created and when it becomes available.